refactor(discovery): split db/src/discovery.rs into authors/series/tags submodules
File was 1210 lines; splits along the natural taxonomy boundary already
named in #308. Each submodule has a short //!. No behavior changes.
Doc trimming (#303), sqlx::Error wrap (#305), ACL plumbing (#232)
deferred.

Closes #308
